Remove the need to change Cinnamon Effect settings
Use _shouldAnimate() patching to prevent Cinnamon from animating the Minimize and Unminimize events. This means that users no longer need to disable the Effect settings in Cinnamon manually. Also means that the Cinnamon New Window animation can be enabled and active along with Magic Lap effects.

Since the CinnamonBurnMyWindows effect also patches _shouldAnimate() in the same way that CinnamonMagicLap now does, we need a way to coordinate the patching so that the two extensions can coexist. This change will introduce a ShouldAnimateManager class that will be used in both extensions. This class will manage the patching of _shouldAnimate() and allow both extensions to intercept different calling contexts of _shouldAnimate().

Updated the Cinnamon requirements from 5.4 to 5.6 since this extension really needs Cinnamon 5.6 to operate correctly.

Some readme changes.