Fix for ShouldAnimateManager
The ShouldAnimateManager was failing to call the original _shouldAnimate() function for cases where the event context was not one of the four types that the manager is currently interested in. This meant that some events like maximize and restore (and maybe others) were getting a return of 'false' in some cases when it should have been 'true' and this results in the cinnamon animations not occurring.